Add Esteban's corrections
My brother suggested to me that the text can be improved by removing the first
person writting style.

I think that for literate programming in particular it's better to describe what
gets done, not what *I*/*we*/*you*/*us* are doing in the code. So, soon these
programs will be revised.